fix timing by using Rune SDK correctly
